fix(spec,core): give pluralToSingular a schema-free home on /meta-spelling (#11503)
Restores the #10096 standing invariant for core's store-key fold: the defineStack manifest-collection vocabulary moves to a leaf module on the /meta-spelling entry graph, /shared re-exports it, and core imports the fold from /meta-spelling instead of /shared.
